requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
python-dotenv
shortuuid
Werkzeug